refactor:consolidation update existing
